# dist/api-docs.json
{
  "swagger": "2.0",
  "basePath": "/api",
...
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "paths": {
...
    "/dashboard/list?{pageIndex}&{pageSize}/{propertyName}&{propertyValue}": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get list of user visit locations",
        "description": "visit locations",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "pageIndex",
            "in": "query",
            "description": "page index",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "page size",
            "required": true,
            "type": "integer"
          },
          {
            "name": "propertyName",
            "in": "query",
            "enum": ["warehouseId"],
            "description": "PropertyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "propertyValue",
            "in": "query",
            "description": "propertyValue",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users Visit response",
            "schema": {
              "$ref": "#/definitions/userVisitLocationPaginationResponse"
            }
          }
        }
      }
    },
    "/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Dashboard ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dash board response",
            "schema": {
              "$ref": "#/definitions/dashboardResponse"
            }
          }
        }
      }
    },
    "/dashboard/visit": {
      "get": {
        "tags": ["Dashboard"],
        "description": "Dashboard ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dash board response",
            "schema": {
              "$ref": "#/definitions/userVisitLocationResponse"
            }
          }
        }
      }
    },
    "/users?{propertyName}&{propertyValue}": {
      "get": {
        "tags": ["User"],
        "description": "Get All users ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "propertyName",
            "in": "query",
            "enum": [
              "userName",
              "firstName",
              "lastName",
              "email",
              "designation",
              "mobileNumber"
            ],
            "description": "PropertyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "propertyValue",
            "in": "query",
            "description": "propertyValue",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/userResponse"
            }
          }
        }
      }
    },
...
    "/users/locations?{searchBy}&{searchByValue}": {
      "get": {
        "tags": ["User"],
        "summary": "Get logged in manager users locations",
        "description": "only accessible to manager role, when searchBy is selected, search by value required else 400 will be thrown",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "searchBy",
            "in": "query",
            "enum": ["firstName", "lastName"],
            "description": "searchBy",
            "required": false,
            "type": "string"
          },
          {
            "name": "searchByValue",
            "in": "query",
            "description": "searchByValue",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/userLocations"
            }
          }
        }
      }
    },
    "/users/{userId}/locations": {
      "get": {
        "tags": ["User"],
        "summary": "Get logged in manager selected user locations of the day",
        "description": "only accessible to manager role, if user is not found 404 will be thrown",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "user Id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/selectedUserLocations"
            }
          }
        }
      }
    },
    "/users/{userId}/last/location": {
      "get": {
        "tags": ["User"],
        "summary": "Get logged in manager selected user last location",
        "description": "only accessible to manager role, if user is not found or last location is not found 404 will be thrown",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "user Id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/lastLocation"
            }
          }
        }
      }
    },
...
    "/users/{userId}": {
      "put": {
        "tags": ["User"],
        "summary": "update user",
        "description": "only accessible to Admin",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users update response",
            "schema": {
              "$ref": "#/definitions/successMessageResponse"
            }
          }
        }
      }
    },
...
    "/users/{type}": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users if user is admin else gets users under manager",
        "description": "only accessible to manager,admin roles",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "type",
            "in": "path",
            "description": "",
            "type": "string",
            "enum": ["All", "ZonalManager", "RegionalManager", "SalesAgent"],
            "required": true
          },
          {
            "name": "searchText",
            "in": "query",
            "description": "",
            "type": "string",
            "required": false
          },
          {
            "name": "userStatus",
            "in": "query",
            "description": "",
            "type": "string",
            "required": false,
            "enum": ["All", "Active", "Deactive"]
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
    "/user/visit/locations": {
      "post": {
        "tags": ["User Visit Location"],
        "description": "User Visit Location",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "user visit location",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userVisitLocation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Visit Location",
            "schema": {
              "$ref": "#/definitions/userVisitLocationData"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "successMessageResponse": {
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "dashboardResponse": {
      "properties": {
        "userCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "customerCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "companyCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "productCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "warehouseCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "preOrderCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "approvedPreOrderCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "pendingPreOrderCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "orderCount": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "offerCount": {
          "required": true,
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "user": {
      "properties": {
        "id": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "name": {
          "required": true,
          "type": "string"
        },
        "firstName": {
          "required": true,
          "type": "string"
        },
        "lastName": {
          "required": true,
          "type": "string",
          "example": "abc@gmail.com"
        },
        "email": {
          "required": true,
          "type": "string"
        },
        "userRole": {
          "required": true,
          "type": "string"
        },
        "designation": {
          "required": true,
          "type": "string"
        },
        "avatar": {
          "required": true,
          "type": "string"
        },
        "mobileNumber": {
          "required": true,
          "type": "string"
        },
        "target": {
          "required": false,
          "type": "number"
        },
        "supervisior": {
          "required": false,
          "type": "string"
        },
        "dateOfJoin": {
          "required": true,
          "type": "string",
          "example": "2019-02-01"
        },
        "employeeId": {
          "required": true,
          "type": "string",
          "example": ""
        },
        "isActive": {
          "required": true,
          "type": "boolean",
          "example": true
        },
        "deactivatedDate": {
          "required": false,
          "type": "string",
          "example": "2018-01-10"
        },
        "hasAgents": {
          "required": true,
          "type": "boolean",
          "example": true
        },
        "warehouseId": {
          "required": false,
          "type": "integer",
          "example": 1
        },
        "warehouseName": {
          "required": false,
          "type": "string",
          "example": "warehouseName"
        },
        "branchId": {
          "required": false,
          "type": "integer",
          "example": 1
        },
        "branchName": {
          "required": false,
          "type": "string",
          "example": "branchName"
        },
        "outletIds": {
          "required": true,
          "type": "array",
          "example": [1]
        }
      }
    },
...
    "userResponse": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/user"
      }
    },
...
    "usersLocation": {
      "properties": {
        "userId": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        },
        "createdAt": {
          "required": true,
          "type": "string",
          "example": "2019-02-02T10:18:16.397Z"
        }
      }
    },
    "userVisitLocation": {
      "properties": {
        "warehouseId": {
          "type": "integer",
          "required": true,
          "example": 1
        },
        "visit": {
          "required": true,
          "type": "boolean",
          "example": true
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        }
      }
    },
    "userVisitLocationData": {
      "properties": {
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "userId": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "warehouseId": {
          "type": "integer",
          "required": true,
          "example": 1
        },
        "visit": {
          "required": true,
          "type": "boolean",
          "example": true
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        },
        "createdAt": {
          "required": true,
          "type": "string",
          "example": "2019-02-02T10:18:16.397Z"
        }
      }
    },
    "userVisitLocationResult": {
      "properties": {
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "userId": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "salesAgentName": {
          "required": true,
          "type": "string",
          "example": ""
        },
        "warehouseId": {
          "type": "integer",
          "required": true,
          "example": 1
        },
        "outletName": {
          "type": "string",
          "required": true,
          "example": ""
        },
        "visit": {
          "required": true,
          "type": "boolean",
          "example": true
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        },
        "createdAt": {
          "required": true,
          "type": "string",
          "example": "2019-02-02T10:18:16.397Z"
        }
      }
    },
    "userVisitLocationPaginationResponse": {
      "properties": {
        "count": {
          "type": "integer",
          "required": true
        },
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/userVisitLocationResult"
          }
        }
      }
    },
    "userVisitLocationResponse": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/userVisitLocationResult"
      }
    },
...
    "Users": {
      "type": "array",
      "items": {
        "properties": {
          "id": {
            "required": true,
            "type": "integer"
          },
          "firstName": {
            "required": true,
            "type": "string"
          },
          "lastName": {
            "required": true,
            "type": "string"
          },
          "userRole": {
            "required": true,
            "type": "string"
          }
        }
      }
    },
    "userUpdate": {
      "properties": {
        "firstName": {
          "type": "string",
          "required": true
        },
        "lastName": {
          "type": "string",
          "required": true
        },
        "email": {
          "type": "string",
          "required": true
        },
        "userName": {
          "type": "string",
          "required": true
        },
        "employeeId": {
          "type": "string",
          "required": true
        },
        "mobileNumber": {
          "type": "string",
          "required": true
        },
        "dateOfJoin": {
          "type": "string",
          "required": true
        }
      }
    },
...
    "userLocations": {
      "type": "array",
      "items": {
        "properties": {
          "id": {
            "required": true,
            "type": "number",
            "example": 1
          },
          "firstName": {
            "required": true,
            "type": "string",
            "example": "john"
          },
          "lastName": {
            "required": true,
            "type": "string",
            "example": "john"
          },
          "avatar": {
            "required": true,
            "type": "string",
            "example": ""
          },
          "email": {
            "required": true,
            "type": "string",
            "example": "john"
          },
          "userRole": {
            "required": true,
            "type": "string",
            "example": ""
          },
          "locations": {
            "required": true,
            "type": "array",
            "items": {
              "$ref": "#/definitions/usersLocation"
            }
          }
        }
      }
    },
...
    "selectedUserLocations": {
      "properties": {
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "firstName": {
          "required": true,
          "type": "string",
          "example": "john"
        },
        "lastName": {
          "required": true,
          "type": "string",
          "example": "john"
        },
        "avatar": {
          "required": true,
          "type": "string",
          "example": ""
        },
        "email": {
          "required": true,
          "type": "string",
          "example": "john"
        },
        "userRole": {
          "required": true,
          "type": "string",
          "example": ""
        },
        "locations": {
          "required": true,
          "type": "array",
          "items": {
            "$ref": "#/definitions/usersLocation"
          }
        }
      }
    },
    "lastLocation": {
      "properties": {
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        },
        "createdAt": {
          "required": true,
          "type": "string",
          "example": "2019-02-02T10:18:16.397Z"
        }
      }
    },
...
  }
}